feat: rename ICoins to ICoin
