+++ backend/src/interfaces/ICoin.ts
